+++ src/enums/serviceMap.ts
@@ -79,15 +79,19 @@ import AzureCosmosDb from '../services/cosmosDb'
 import AzureAppInsights from '../services/appInsights'
 import AzureIntegrationRuntime from '../services/integrationRuntimes' 
 import AzureServiceBus from '../services/serviceBus'
-import AzureBackupVault from '../services/backupVault'
+// import AzureBackupVault from '../services/backupVault'
 import AzureBackupInstance from '../services/backupInstance'
 import AzureBackupPolicy from '../services/backupPolicy'
+import AzureBilling from '../services/billing'
+import AzureLogProfiles from '../services/logProfiles'
+import Subscription from '../services/subscription'
 
 /**
  * serviceMap is an object that contains all currently supported services for AWS
  * serviceMap is used by the serviceFactory to produce instances of service classes
  */
 export default {
+  subscription: Subscription,
   [services.actionGroup]: AzureActionGroup,
   [services.activityLogAlerts]: AzureActivityLogAlerts,
   [services.adApplication]: AzureADApplication,
@@ -107,12 +111,12 @@ export default {
   [services.authRoleAssignment]: AzureAuthRoleAssignment,
   [services.authRoleDefinition]: AzureAuthRoleDefinition,
   [services.autoProvisioningSettings]: AzureAutoProvisioningSettings,
-  [services.backupVault]: AzureBackupVault,
+  // [services.backupVault]: AzureBackupVault,
   [services.backupInstance]: AzureBackupInstance,
   [services.backupPolicy]: AzureBackupPolicy,
+  [services.billing]: AzureBilling,
   [services.cdnCustomDomains]: AzureCdnCustomDomains,
   [services.cdnEndpoints]: AzureCdnEndpoints,
-  [services.cdnOriginGroups]: AzureCdnOriginGroups,
   [services.cdnOriginGroups]: AzureCdnOriginGroups,
   [services.cdnOrigins]: AzureCdnOrigins,
   [services.cdnOrigins]: AzureCdnOrigins,
@@ -142,6 +146,7 @@ export default {
   [services.loadBalancer]: AzureLoadBalancer,
   [services.logAnalyticsSolution]: AzureLogAnalyticsSolutions,
   [services.logAnalyticsWorkspace]: AzureLogAnalyticsWorkspaces,
+  [services.logProfiles]: AzureLogProfiles,
   [services.machineLearningWorkspaces]: AzureMachineLearningWorkspace,
   [services.mySqlServers]: AzureMySqlServer,
   [services.networkInterface]: AzureNetworkInterface,