+++ src/services/diagnosticSettings/data.ts
@@ -3,7 +3,7 @@ import { MonitorClient, DiagnosticSettingsResource } from '@azure/arm-monitor'
 
 import azureLoggerText from '../../properties/logger'
 import { lowerCaseLocation } from '../../utils/format'
-import { AzureServiceInput } from '../../types'
+import { AzureServiceConfig, AzureServiceInput } from '../../types'
 import { getResourceGroupFromEntity } from '../../utils/idParserUtils'
 import { tryCatchWrapper } from '../../utils'
 
@@ -17,50 +17,61 @@ export interface RawAzureDiagnosticSetting
   resourceGroupId: string
 }
 
+export const getDiagnosticSettings = async (
+  config: AzureServiceConfig,
+  resourceUri: string
+): Promise<DiagnosticSettingsResource[]> => {
+  const { tokenCredentials, subscriptionId } = config
+  const client = new MonitorClient(tokenCredentials, subscriptionId)
+  const diagnosticSettings: DiagnosticSettingsResource[] = []
+  await tryCatchWrapper(
+    async () => {
+      const { value: settings = [] } = await client.diagnosticSettings.list(
+        resourceUri
+      )
+      if (settings) {
+        diagnosticSettings.push(...settings)
+      }
+    },
+    {
+      service: serviceName,
+      client,
+      scope: 'diagnosticSettings',
+      operation: 'list',
+    }
+  )
+  logger.debug(lt.foundDiagnosticSettingsResources(diagnosticSettings.length))
+  return diagnosticSettings
+}
+
 export default async ({
   regions,
   config,
 }: AzureServiceInput): Promise<{
   [property: string]: RawAzureDiagnosticSetting[]
 }> => {
-  const { tokenCredentials, subscriptionId } = config
-  const client = new MonitorClient(tokenCredentials, subscriptionId)
-
+  const { subscriptionId } = config
   try {
     const result: { [property: string]: RawAzureDiagnosticSetting[] } = {}
-    let numOfGroups = 0
-    await tryCatchWrapper(
-      async () => {
-        const diagnosticSettings: DiagnosticSettingsResource[] =
-          (
-            await client.diagnosticSettings.list(
-              `/subscriptions/${subscriptionId}`
-            )
-          )?.value || []
-        numOfGroups = diagnosticSettings.length
-        diagnosticSettings.map(d => {
-          const region = lowerCaseLocation('')
-          if (regions.includes(region)) {
-            if (!result[region]) {
-              result[region] = []
-            }
-            const resourceGroupId = getResourceGroupFromEntity(d)
-            result[region].push({
-              ...d,
-              resourceGroupId,
-              region,
-            })
-          }
-        })
-      },
-      {
-        service: serviceName,
-        client,
-        scope: 'diagnosticSettings',
-        operation: 'list',
-      }
+    const diagnosticSettings = await getDiagnosticSettings(
+      config,
+      `/subscriptions/${subscriptionId}`
     )
-    logger.debug(lt.foundDiagnosticSettingsResources(numOfGroups))
+    diagnosticSettings.map(d => {
+      const region = lowerCaseLocation('')
+      if (regions.includes(region)) {
+        if (!result[region]) {
+          result[region] = []
+        }
+        const resourceGroupId = getResourceGroupFromEntity(d)
+        result[region].push({
+          ...d,
+          resourceGroupId,
+          region,
+        })
+      }
+    })
+
     return result
   } catch (e) {
     logger.debug(e)