+++ src/services/expressRouteGateways/format.ts
@@ -24,22 +24,27 @@ export default ({
     virtualHub,
     Tags = {},
   } = service
-
   return {
     id: id || cuid(),
     name,
     type,
     region,
+    subscriptionId: account,
     etag,
     autoScaleConfiguration,
     expressRouteConnections: expressRouteConnections?.map(connection => ({
-      ...connection,
-      id: id || cuid(),
+      id: cuid(),
+      connectionId: id,
+      name: connection?.name,
+      provisioningState: connection?.provisioningState,
+      authorizationKey: connection?.authorizationKey,
+      routingWeight: connection?.routingWeight,
+      enableInternetSecurity: connection?.enableInternetSecurity,
+      expressRouteGatewayBypass: connection?.expressRouteGatewayBypass,
       expressRouteCircuitPeering: {
         id: connection?.expressRouteCircuitPeering?.id,
       },
       routingConfiguration: {
-        ...connection?.routingConfiguration,
         associatedRouteTable: {
           id: connection?.routingConfiguration?.associatedRouteTable?.id || cuid(),
         },